update README, changed int format to i64 for best results (just buy more ram\!)

+++ tests/tests.odin
@@ -24,7 +24,7 @@ simple :: proc(t: ^testing.T) {
 	testing.expect(t, err == .None)
 	testing.expect(t, len(data.super_table["list"].(toml.Toml_Array)) == 9)
 	testing.expect(t, data.super_table["bool"].(bool) == true)
-	testing.expect(t, data.super_table["other"].(toml.Toml_Map)["hi"].(int) == 1323)
+	testing.expect(t, data.super_table["other"].(toml.Toml_Map)["hi"].(i64) == 1323)
 	testing.expect(t, data.super_table["string"].(string) == "value")
 	testing.expect(t, data.super_table["float"].(f64) - 82.233333333333337 < 0.01)
 	testing.expect(t, data.super_table["inline"].(toml.Toml_Map)["key"].(string) == inline["key"].(string))
